Moved unix testing to new file
